[iOS] style: 모델 코드 분리
